+++ src/game/tests.rs
@@ -428,6 +428,7 @@ mod vintage_cube_forth_eorlingas;
 mod vintage_cube_fractured_identity;
 mod vintage_cube_gadgeteer;
 mod vintage_cube_galvanic_blast;
+mod vintage_cube_galvanic_discharge;
 mod vintage_cube_gau;
 mod vintage_cube_generous_plunderer;
 mod vintage_cube_get_lost;